move out scalers from detectors
